+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/pnl/GrossLossCriterion.java
@@ -39,25 +39,18 @@ import org.ta4j.core.num.Num;
 public class GrossLossCriterion extends AbstractAnalysisCriterion {
 
     @Override
-    public Num calculate(BarSeries series, TradingRecord tradingRecord) {
-        return tradingRecord.getPositions().stream().filter(Position::isClosed)
-                .map(position -> calculate(series, position)).reduce(series.numOf(0), Num::plus);
-    }
-
-    /**
-     * Calculates the gross loss value of given position
-     *
-     * @param series   a bar series
-     * @param position a position to calculate profit
-     * @return the gross loss
-     */
-    @Override
     public Num calculate(BarSeries series, Position position) {
         if (position.isClosed()) {
             Num loss = position.getGrossProfit();
             return loss.isNegative() ? loss : series.numOf(0);
         }
         return series.numOf(0);
+    }
+
+    @Override
+    public Num calculate(BarSeries series, TradingRecord tradingRecord) {
+        return tradingRecord.getPositions().stream().filter(Position::isClosed)
+                .map(position -> calculate(series, position)).reduce(series.numOf(0), Num::plus);
     }
 
     @Override