+++ ta4j-core/src/main/java/org/ta4j/core/reports/PerformanceReport.java
@@ -57,5 +57,4 @@ public class PerformanceReport {
     public Num getTotalLoss() {
         return totalLoss;
     }
-
 }