+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/ExpectancyCriterionTest.java
@@ -21,7 +21,7 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.analysis.criteria.pnl;
+package org.ta4j.core.analysis.criteria;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -34,44 +34,63 @@ import org.ta4j.core.AnalysisCriterion;
 import org.ta4j.core.BaseTradingRecord;
 import org.ta4j.core.Trade;
 import org.ta4j.core.TradingRecord;
-import org.ta4j.core.analysis.criteria.AbstractCriterionTest;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
 
-public class NetLossCriterionTest extends AbstractCriterionTest {
+public class ExpectancyCriterionTest extends AbstractCriterionTest {
 
-    public NetLossCriterionTest(Function<Number, Num> numFunction) {
-        super((params) -> new NetLossCriterion(), numFunction);
+    public ExpectancyCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new ExpectancyCriterion(), numFunction);
     }
 
     @Test
     public void calculateOnlyWithProfitPositions() {
-        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 110, 100, 95, 105);
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 110, 120, 130, 150, 160);
         TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(2, series),
                 Trade.buyAt(3, series), Trade.sellAt(5, series));
 
-        AnalysisCriterion loss = getCriterion();
-        assertNumEquals(0, loss.calculate(series, tradingRecord));
+        AnalysisCriterion avgLoss = getCriterion();
+        assertNumEquals(0, avgLoss.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateWithMixedPositions() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 110, 80, 130, 150, 160);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(2, series),
+                Trade.buyAt(3, series), Trade.sellAt(5, series));
+
+        AnalysisCriterion avgLoss = getCriterion();
+        assertNumEquals(-1.25, avgLoss.calculate(series, tradingRecord));
     }
 
     @Test
     public void calculateOnlyWithLossPositions() {
-        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 80, 70, 60, 50);
         TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series),
                 Trade.buyAt(2, series), Trade.sellAt(5, series));
 
-        AnalysisCriterion loss = getCriterion();
-        assertNumEquals(-35, loss.calculate(series, tradingRecord));
+        AnalysisCriterion avgLoss = getCriterion();
+        assertNumEquals(0, avgLoss.calculate(series, tradingRecord));
     }
 
     @Test
     public void calculateProfitWithShortPositions() {
-        MockBarSeries series = new MockBarSeries(numFunction, 95, 100, 70, 80, 85, 100);
+        MockBarSeries series = new MockBarSeries(numFunction, 160, 140, 120, 100, 80, 60);
         TradingRecord tradingRecord = new BaseTradingRecord(Trade.sellAt(0, series), Trade.buyAt(1, series),
                 Trade.sellAt(2, series), Trade.buyAt(5, series));
 
-        AnalysisCriterion loss = getCriterion();
-        assertNumEquals(-35, loss.calculate(series, tradingRecord));
+        AnalysisCriterion avgLoss = getCriterion();
+        assertNumEquals(0, avgLoss.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateProfitWithMixedShortPositions() {
+        MockBarSeries series = new MockBarSeries(numFunction, 160, 200, 120, 100, 80, 60);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.sellAt(0, series), Trade.buyAt(1, series),
+                Trade.sellAt(2, series), Trade.buyAt(5, series));
+
+        AnalysisCriterion avgLoss = getCriterion();
+        assertNumEquals(-1.25, avgLoss.calculate(series, tradingRecord));
     }
 
     @Test
@@ -85,4 +104,5 @@ public class NetLossCriterionTest extends AbstractCriterionTest {
     public void testCalculateOneOpenPositionShouldReturnZero() {
         openedPositionUtils.testCalculateOneOpenPositionShouldReturnExpectedValue(numFunction, getCriterion(), 0);
     }
+
 }