+++ ta4j-core/src/test/java/org/ta4j/core/indicators/DistanceFromMAIndicatorTest.java
@@ -25,47 +25,40 @@ package org.ta4j.core.indicators;
 
 import static org.ta4j.core.TestUtils.assertNumEquals;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.function.Function;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.ta4j.core.Bar;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
-import org.ta4j.core.mocks.MockBar;
+import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
 
-public class PVOIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
+public class DistanceFromMAIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
+    private BarSeries data;
 
-    private BarSeries barSeries;
-
-    public PVOIndicatorTest(Function<Number, Num> numFunction) {
+    public DistanceFromMAIndicatorTest(Function<Number, Num> numFunction) {
         super(numFunction);
     }
 
     @Before
     public void setUp() {
-        List<Bar> bars = new ArrayList<Bar>();
-        bars.add(new MockBar(0, 10, numFunction));
-        bars.add(new MockBar(0, 11, numFunction));
-        bars.add(new MockBar(0, 12, numFunction));
-        bars.add(new MockBar(0, 13, numFunction));
-        bars.add(new MockBar(0, 150, numFunction));
-        bars.add(new MockBar(0, 155, numFunction));
-        bars.add(new MockBar(0, 160, numFunction));
-        barSeries = new MockBarSeries(bars);
+        data = new MockBarSeries(numFunction, 10, 15, 20, 18, 17, 18, 15, 12, 10, 8, 5, 2);
     }
 
     @Test
-    public void createPvoIndicator() {
-        PVOIndicator pvo = new PVOIndicator(barSeries);
-        assertNumEquals(0.791855204, pvo.getValue(1));
-        assertNumEquals(2.164434756, pvo.getValue(2));
-        assertNumEquals(3.925400464, pvo.getValue(3));
-        assertNumEquals(55.296120101, pvo.getValue(4));
-        assertNumEquals(66.511722064, pvo.getValue(5));
+    public void DistanceFromMovingAverageTest() {
+        SMAIndicator sma = new SMAIndicator(new ClosePriceIndicator(data), 3);
+        DistanceFromMAIndicator distanceFromMAIndicator = new DistanceFromMAIndicator(data, sma);
+        assertNumEquals(0.3333, distanceFromMAIndicator.getValue(2));
+        assertNumEquals(0.01886792452830182, distanceFromMAIndicator.getValue(5));
+        assertNumEquals(-0.1, distanceFromMAIndicator.getValue(6));
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void DistanceFromIllegalMovingAverage() {
+        ClosePriceIndicator closePriceIndicator = new ClosePriceIndicator(data);
+        new DistanceFromMAIndicator(data, closePriceIndicator);
     }
 }